fix: api 폴더 다시 변경
