+++ src/apis/templeDetail/type.ts
@@ -1,0 +1,17 @@
+export interface TempleDetail {
+  templestayId: string;
+  templeName: string;
+  templestayName: string;
+  phoneNumber: string;
+  address: string;
+  tag?: string;
+  templestayPrice?: string;
+  introduction?: string;
+  detailAddress: string;
+  youtube?: string;
+  schedule?: string;
+  latitude: number;
+  longitude: number;
+  liked: boolean;
+  url: string;
+}
