+++ src/apis/templeImages/axios.ts
@@ -1,0 +1,10 @@
+import instance from '@apis/instance';
+
+const getTempleImages = async (templestayId: string) => {
+  const res = await instance.get('/public/templestay/img', {
+    params: { templestayId },
+  });
+  return res.data;
+};
+
+export default getTempleImages;
