+++ src/apis/templeImages/type.ts
@@ -1,0 +1,9 @@
+export interface TemplestayImg {
+  imageUrlId: number;
+  imgUrl: string;
+}
+
+export interface TemplestayImgsResponse {
+  total: number;
+  templestayImgs: TemplestayImg[];
+}
