+++ src/apis/templeReviews/type.ts
@@ -1,0 +1,18 @@
+export interface Review {
+  reviewId?: string;
+  reviewTitle?: string;
+  reviewLink?: string;
+  reviewName?: string;
+  reviewDescription?: string;
+  reviewDate?: string;
+  reviewImgUrl?: string;
+}
+
+export interface ReviewsResponse {
+  templestayId: string;
+  page: number;
+  pageSize: number;
+  totalPages: number;
+  reviewCount: number;
+  reviews?: Review[];
+}
