+++ src/pages/templeDetailPage/templePhoto/TemplePhotoPage.tsx
@@ -1,4 +1,4 @@
-import { useGetTempleImages } from '@apis/templeInfo';
+import useGetTempleImages from '@apis/templeImages';
 import PageName from '@components/common/pageName/PageName';
 import ExceptLayout from '@components/except/exceptLayout/ExceptLayout';
 import { useParams } from 'react-router-dom';
